refactor: move wrapper and partial factories into separate files
