+++ src/prefabs/factories/component.ts
@@ -1,7 +1,5 @@
 import type { IdentityRecordBy } from '../../type-utils';
-import { LinkedOptionProducer } from '../types';
 import type { PrefabComponent, PrefabReference } from '../types/component';
-import type { OptionCategory } from '../types/options';
 
 function isNotNullEntry<T, X>(entry: [T, X]): entry is [T, NonNullable<X>] {
   const [, option] = entry;
@@ -41,39 +39,6 @@ export const partial = (): PrefabReference => ({
   type: 'PARTIAL',
   partialId: '',
 });
-
-export type WrapperAttrs = {
-  label?: string;
-  optionCategories?: OptionCategory[];
-  options?: Record<string, LinkedOptionProducer>;
-};
-
-/**
- * Create a wrapper prefab
- *
- * @returns
- */
-export const wrapper = (
-  attrs: WrapperAttrs,
-  descendants: PrefabReference[],
-): PrefabReference => {
-  const labelField = attrs.label ? { label: attrs.label } : {};
-  const options = Object.entries(attrs.options || {}).map(([key, linked]) =>
-    linked(key),
-  );
-  const optionCategories =
-    attrs.optionCategories && attrs.optionCategories.length !== 0
-      ? { optionCategories: attrs.optionCategories }
-      : {};
-
-  return {
-    type: 'WRAPPER',
-    ...labelField,
-    ...optionCategories,
-    options,
-    descendants,
-  };
-};
 
 /**
  * Create a component prefab
