+++ src/prefabs/factories/index.ts
@@ -1,3 +1,5 @@
-export { component, wrapper, partial } from './component';
+export { component } from './component';
+export { wrapper } from './wrapper';
+export { partial } from './partial';
 export { prefab, BeforeCreateArgs } from './prefab';
 export * from './options';
