+++ src/prefabs/factories/partial.ts
@@ -1,0 +1,11 @@
+import type { PrefabReference } from '../types/component';
+
+/**
+ * Create a partial prefab
+ *
+ * @returns
+ */
+export const partial = (): PrefabReference => ({
+  type: 'PARTIAL',
+  partialId: '',
+});
