+++ tests/prefabs/factories/partial.test.ts
@@ -1,0 +1,14 @@
+import test from 'tape';
+
+import { partial } from '../../../src/prefabs/factories/partial';
+
+test('partial builds empty partial', (t) => {
+  const result = partial();
+  const expected = {
+    type: 'PARTIAL',
+    partialId: '',
+  };
+
+  t.deepEqual(result, expected);
+  t.end();
+});
